# backend/src/openapi-generated.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Stellar MicroPay API",
    "version": "1.0.0",
    "description": "Backend API for Stellar MicroPay — instant micropayments on the Stellar network.\n\n## Rate Limiting\n\nAll endpoints are rate-limited. Three limiters apply:\n\n| Limiter | Window | Limit | Routes |\n|---------|--------|-------|--------|\n| Global | 15 minutes | 100 req/IP | All routes |\n| Strict | 1 minute | 20 req/IP | `/api/analytics/*`, `/api/tips/*`, `/api/webhooks/*`, `/federation`, `/api/accounts/register` |\n| Payment | 1 minute | 10 req/IP | `/api/payments/*`, `/api/turrets/*` |\n\nEvery response includes the following headers so clients can implement back-off:\n\n| Header | Description |\n|--------|-------------|\n| `RateLimit-Limit` | Maximum requests allowed in the current window |\n| `RateLimit-Remaining` | Requests remaining before the limit is reached |\n| `RateLimit-Reset` | Seconds until the window resets |\n\nWhen the limit is exceeded the server returns **HTTP 429** with `{ \"error\": \"Too many requests, please try again later.\" }`. Clients should read `RateLimit-Remaining` on each response and add exponential back-off when the value approaches 0.",
    "contact": {
      "name": "Stellar MicroPay",
      "url": "https://github.com/Emmy123222/Stellar-MicroPay"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/api/analytics/{publicKey}/summary": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "summary": "Get payment summary for an account",
        "description": "Aggregates the account's most recent payments (up to 200) into totals sent and received, unique counterparties, average transaction size, and a week-over-week comparison. Results are cached for 60 seconds.",
        "parameters": [
          {
            "name": "publicKey",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^G[A-Z0-9]{55}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Analytics summary",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "publicKey": {
                          "type": "string"
                        },
                        "totalSentXLM": {
                          "type": "string"
                        },
                        "totalReceivedXLM": {
                          "type": "string"
                        },
                        "uniqueCounterparties": {
                          "type": "integer"
                        },
                        "averageTransactionSize": {
                          "type": "string"
                        },
                        "totalTransactions": {
                          "type": "integer"
                        },
                        "comparison": {
                          "type": "object",
                          "description": "Week-over-week comparison of payment count and volume",
                          "properties": {
                            "thisWeekCount": {
                              "type": "integer"
                            },
                            "lastWeekCount": {
                              "type": "integer"
                            },
                            "countChangePercent": {
                              "type": "integer"
                            },
                            "thisWeekVolume": {
                              "type": "string"
                            },
                            "lastWeekVolume": {
                              "type": "string"
                            },
                            "volumeChangePercent": {
                              "type": "integer"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid Stellar public key format"
          },
          "429": {
            "description": "Rate limit exceeded"
          }
        }
      }
    },
    "/api/analytics/{publicKey}/top-recipients": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "summary": "Get top payment recipients",
        "description": "Ranks the top 5 recipients by total XLM sent, counting sent payments from the account's most recent payments (up to 200), sorted descending by total. Results are cached for 60 seconds.",
        "parameters": [
          {
            "name": "publicKey",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^G[A-Z0-9]{55}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Top recipients by total XLM sent",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "publicKey": {
                          "type": "string"
                        },
                        "topRecipients": {
                          "type": "array",
                          "description": "Up to 5 entries, highest total first.",
                          "items": {
                            "type": "object",
                            "properties": {
                              "address": {
                                "type": "string",
                                "description": "Recipient Stellar public key"
                              },
                              "totalXLMSent": {
                                "type": "string",
                                "description": "Total XLM sent to the recipient"
                              }
                            }
                          }
                        },
                        "count": {
                          "type": "integer",
                          "description": "Number of recipients returned."
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid Stellar public key format"
          },
          "429": {
            "description": "Rate limit exceeded"
          }
        }
      }
    },
    "/api/analytics/{publicKey}/activity": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "summary": "Get payment activity by day of week",
        "description": "Counts the account's most recent payments (up to 200) by day of week, always returning all seven days (Sunday first). Results are cached for 60 seconds.",
        "parameters": [
          {
            "name": "publicKey",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^G[A-Z0-9]{55}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Payment counts for each day of the week",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "publicKey": {
                          "type": "string"
                        },
                        "activityByDay": {
                          "type": "array",
                          "description": "Always 7 entries, Sunday through Saturday.",
                          "items": {
                            "type": "object",
                            "properties": {
                              "day": {
                                "type": "string",
                                "description": "Day of week name",
                                "example": "Monday"
                              },
                              "dayIndex": {
                                "type": "integer",
                                "description": "Day of week index (0 = Sunday, 6 = Saturday)",
                                "minimum": 0,
                                "maximum": 6
                              },
                              "transactionCount": {
                                "type": "integer",
                                "description": "Number of payments on that day"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid Stellar public key format"
          },
          "429": {
            "description": "Rate limit exceeded"
          }
        }
      }
    },
    "/api/analytics/{publicKey}/cohorts": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "summary": "Get repeat-vs-one-time counterparty cohorts",
        "description": "Groups the account's most recent payments (up to 200) into calendar buckets and reports, per bucket, how many counterparties appeared once versus more than once, split by sent and received payments. Results are cached for 60 seconds.",
        "parameters": [
          {
            "name": "publicKey",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^G[A-Z0-9]{55}$"
            }
          },
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "month",
                "week"
              ],
              "default": "month"
            },
            "description": "Bucket size for the cohort breakdown. Defaults to month; any value other than week falls back to month."
          },
          {
            "name": "periods",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 12,
              "default": 6
            },
            "description": "How many buckets to return, ending with the current period. Defaults to 6; values below 1 or that are not numeric fall back to 6, and values above 12 are capped at 12."
          }
        ],
        "responses": {
          "200": {
            "description": "Cohort breakdown",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "publicKey": {
                          "type": "string"
                        },
                        "period": {
                          "type": "string",
                          "enum": [
                            "month",
                            "week"
                          ]
                        },
                        "periods": {
                          "type": "integer"
                        },
                        "range": {
                          "type": "object",
                          "properties": {
                            "start": {
                              "type": "string",
                              "format": "date-time",
                              "nullable": true
                            },
                            "end": {
                              "type": "string",
                              "format": "date-time",
                              "nullable": true
                            }
                          }
                        },
                        "cohorts": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "periodStart": {
                                "type": "string",
                                "format": "date-time"
                              },
                              "periodEnd": {
                                "type": "string",
                                "format": "date-time"
                              },
                              "label": {
                                "type": "string"
                              },
                              "period": {
                                "type": "string",
                                "enum": [
                                  "month",
                                  "week"
                                ]
                              },
                              "sent": {
                                "type": "object",
                                "properties": {
                                  "paymentCount": {
                                    "type": "integer"
                                  },
                                  "totalXLM": {
                                    "type": "string"
                                  },
                                  "counterparties": {
                                    "type": "object",
                                    "properties": {
                                      "oneTimeCounterparties": {
                                        "type": "integer"
                                      },
                                      "repeatCounterparties": {
                                        "type": "integer"
                                      },
                                      "totalCounterparties": {
                                        "type": "integer"
                                      }
                                    }
                                  }
                                }
                              },
                              "received": {
                                "type": "object",
                                "properties": {
                                  "paymentCount": {
                                    "type": "integer"
                                  },
                                  "totalXLM": {
                                    "type": "string"
                                  },
                                  "counterparties": {
                                    "type": "object",
                                    "properties": {
                                      "oneTimeCounterparties": {
                                        "type": "integer"
                                      },
                                      "repeatCounterparties": {
                                        "type": "integer"
                                      },
                                      "totalCounterparties": {
                                        "type": "integer"
                                      }
                                    }
                                  }
                                }
                              },
                              "totalCounterparties": {
                                "type": "integer"
                              },
                              "repeatRate": {
                                "type": "integer"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid Stellar public key format"
          },
          "429": {
            "description": "Rate limit exceeded"
          }
        }
      }
    },
...
    "/api/analytics/{publicKey}/export-trigger": {
      "post": {
        "tags": [
          "Analytics"
        ],
        "summary": "Manually trigger sending export email",
        "description": "Sends the export email immediately using the account's stored schedule. Fails with 404 when no schedule exists for the account.",
        "parameters": [
          {
            "name": "publicKey",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^G[A-Z0-9]{55}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Export triggered",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "success": {
                          "type": "boolean",
                          "example": true
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "Export email sent"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid Stellar public key format"
          },
          "404": {
            "description": "No export schedule found"
          },
          "429": {
            "description": "Rate limit exceeded"
          }
        }
      }
    },
...
  }
}